% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{dmCache}
\alias{dmCache}
\title{Depmap Package Cache}
\usage{
dmCache()
}
\value{
The `dmCache()` function returns an instance of class
    `BiocFileCache`.
}
\description{
Function to access (and manage) the depmap package
cache. `dmCache()` returns the central `depmap` cache.
}
\details{
The cache is an object of class `BiocFileCache`, and created with
[BiocFileCache::BiocFileCache()]. It can be either the
package-wide cache as defined by `dmCache()` or an instaned
provided by the user.

When projects are cached, they are given a resource name
(`rname`), that is set to the file's identifier.
}
\examples{

## Default depmap cache
dmCache()
}
\author{
Laurent Gatto
}
