% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCLE_depMap_19Q1_TPM.R
\docType{data}
\name{CCLE_depMap_19Q1_TPM}
\alias{CCLE_depMap_19Q1_TPM}
\title{CCLE_depMap_19Q1_TPM}
\format{A data frame with 1165 rows (cell lines) and 57821 variables (genes):
\describe{
     \item{Gene}{HUGO symbol (e.g. "HSPA5") and Ensembl gene ID (e.g. ENSG00000044574)}
}}
\source{
\href{"https://depmap.org/portal/download/"}{DepMap, Broad (2019)}
}
\usage{
data(CCLE_depMap_19Q1_TPM)
}
\description{
CCLE RNAseq gene expression data at scale (log2(TPM+1)). The portal shows
expression data for only protein coding genes.
}
\details{
Genes 57820
Cell Lines 1165
Primary Diseases 33
Lineages 32
}
\examples{
data(CCLE_depMap_19Q1_TPM)
ACTC1_(70) <- attr(CCLE_depMap_19Q1_TPM, "ACTC1 (70)")

\dontrun{
CCLE_depMap_19Q1_TPM
}

}
\references{
Tsherniak, A., Vazquez, F., Montgomery, P. G., Weir, B. A.,
Kryukov, G., Cowley, G. S., ... & Meyers, R. M. (2017). Defining a cancer
dependency map. Cell, 170(3), 564-576.
(\href{"https://www.ncbi.nlm.nih.gov/pubmed/28753430"}{PubMed})
}
\keyword{datasets}
