% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCLE_RPPA_20180123.R
\docType{data}
\name{CCLE_RPPA_20180123}
\alias{CCLE_RPPA_20180123}
\title{CCLE_RPPA_20180123}
\format{A data frame with 899 rows (cell lines) and 215 variables (genes):
\describe{
    \item{Gene}{Gene (i.e. "Transglutaminase")}
}}
\source{
\href{"https://depmap.org/portal/download/"}{DepMap, Broad (2019)}
}
\usage{
data(CCLE_RPPA_20180123)
}
\description{
Celluar Models dataset: CCLE Reverse Phase Protein Array (RPPA) data
}
\details{
Genes 214
Cell Lines 899
Primary Diseases 28
Lineages 28
}
\examples{
data(CCLE_RPPA_20180123)
ACTC1_(70) <- attr(CCLE_RPPA_20180123, "ACTC1 (70)")

\dontrun{
CCLE_RPPA_20180123
}

}
\references{
Tsherniak, A., Vazquez, F., Montgomery, P. G., Weir, B. A.,
Kryukov, G., Cowley, G. S., ... & Meyers, R. M. (2017). Defining a cancer
dependency map. Cell, 170(3), 564-576.
(\href{"https://www.ncbi.nlm.nih.gov/pubmed/28753430"}{PubMed})
}
\keyword{datasets}
