% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depmap_19Q1_cell_lines.R
\docType{data}
\name{depmap_19Q1_cell_lines}
\alias{depmap_19Q1_cell_lines}
\title{depmap_19Q1_cell_lines}
\format{A data frame with 1677 rows (cell lines) and 9 variables:
\describe{
  \item{DepMap_ID}{DepMap_ID}
  \item{CCLE_Name}{CCLE_Name}
  \item{Aliases}{Aliases}
  \item{COSMIC_ID}{Catalogue Of Somatic Mutations In Cancer (COSMIC) ID number (e.g. 905933)}
  \item{Sanger ID}{Sanger ID (eg. 2201)}
  \item{Primary Disease}{Primary Disease (e.g. cancer type)}
  \item{Subtype Disease}{Subtype Disease (e.g. Acute Myelogenous Leukemia (AML), M3 (Promyelocytic))}
  \item{Gender}{Gender of patient of cell source}
  \item{Source}{Organization where the cell was sampled}
}}
\source{
\href{"https://depmap.org/portal/download/"}{DepMap, Broad (2019)}
}
\usage{
data(depmap_19Q1_cell_lines)
}
\description{
Metadata about cell lines in the 19Q1 release, including mapping
between DepMap ID and CCLE names
}
\details{
Genes 0
Cell Lines 1677
Primary Diseases 38
Lineages 33
}
\examples{
data(depmap_19Q1_cell_lines)
DepMap_ID <- attr(depmap_19Q1_cell_lines, "DepMap_ID")

\dontrun{
depmap_19Q1_cell_lines
}

}
\references{
Tsherniak, A., Vazquez, F., Montgomery, P. G., Weir, B. A.,
Kryukov, G., Cowley, G. S., ... & Meyers, R. M. (2017). Defining a cancer
dependency map. Cell, 170(3), 564-576.
(\href{"https://www.ncbi.nlm.nih.gov/pubmed/28753430"}{PubMed})
}
\keyword{datasets}
