% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_effect_corrected.R
\docType{data}
\name{gene_effect_corrected}
\alias{gene_effect_corrected}
\title{gene_effect_corrected}
\format{A data frame with 558 rows (cell lines) and 17635 variables (genes):
\describe{
    \item{Gene}{HUGO symbol (e.g. "SAP25") and Entrez ID# (e.g. 100316904)}
}}
\source{
\href{"https://ndownloader.figshare.com/files/14221385"}{DepMap, Broad (2019)}
}
\usage{
data(gene_effect_corrected)
}
\description{
Genetic dependency of CRISPR (Avana) Public 19Q1 from batch corrected CERES
inferred gene effect matrix. CERES is a computational method to estimate gene
dependency levels from CRISPR-Cas9 essentiality screens while accounting for
the copy-number-specific effect.
}
\details{
Genes 17634
Cell Lines 558
Primary Diseases 26
Lineages 28
}
\examples{
data(gene_effect_corrected)
A1CF_(29974) <- attr(gene_effect_corrected, "A1CF (29974)")

\dontrun{
gene_effect_corrected
}

}
\references{
Tsherniak, A., Vazquez, F., Montgomery, P. G., Weir, B. A.,
Kryukov, G., Cowley, G. S., ... & Meyers, R. M. (2017). Defining a cancer
dependency map. Cell, 170(3), 564-576.
(\href{"https://www.ncbi.nlm.nih.gov/pubmed/28753430"}{PubMed})
}
\keyword{datasets}
