% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_19Q1.R, R/metadata_19Q2.R
\docType{data}
\name{metadata}
\alias{metadata}
\alias{metadata_19Q1}
\alias{metadata_19Q2}
\title{metadata_19Q1}
\format{A data frame with 1677 rows (cell lines) and 9 variables:
\describe{
  \item{depmap_id}{Cancer cell line primary key, used in other datasets as
  foreign key (i.e. "ACH-00001")}
  \item{cell_line}{CCLE name of cancer cell line (i.e. "184A1_BREAST")}
  \item{aliases}{Aliases of cancer cell line}
  \item{cosmic_id}{Catalogue Of Somatic Mutations In Cancer (COSMIC) ID
  number (e.g. 905933)}
  \item{sanger_id}{Sanger ID (eg. 2201)}
  \item{primary_disease}{Primary Disease (e.g. cancer type)}
  \item{subtype_disease}{Subtype Disease (e.g. Acute Myelogenous Leukemia
  (AML), M3 (Promyelocytic))}
  \item{gender}{Gender of patient of cell source}
  \item{source}{Organization responsible for cell sampling}
}}
\source{
\href{"https://depmap.org/portal/download/"}{DepMap, Broad (2019)}

\href{"https://depmap.org/portal/download/"}{DepMap, Broad (2019)}
}
\usage{
metadata

metadata
}
\description{
The `metadata` dataset contains the metadata about cell lines in the 19Q1
Broad Institute DepMap release, which includes mapping between `depmap_id`
and `cell_line` name for cancer cell lines. This dataset does not contain any
dependency data but contains the metadata for 0 genes, 1677 cell lines, 38
primary diseases, 33 lineages. The columns of `metadata` are: `depmap_id`,
`cell_line`, `aliases`, `cosmic_id`, `sanger_id`, `primary_disease`,
`subtype_disease`, `gender` and `source`.

The `metadata` dataset contains the metadata about cell lines in the 19Q2
Broad Institute DepMap release, which includes mapping between `depmap_id`
and `cell_line` name for cancer cell lines. This dataset does not contain any
dependency data but contains the metadata for 0 genes, 1714 cell lines, 39
primary diseases, 34 lineages. The columns of `metadata` are: `depmap_id`,
`stripped_cell_line_name`, `cell_line`, `aliases`, `cosmic_id`, `sanger_id`,
`primary_disease`, `subtype_disease`, `sub_subtype_disease`, `gender`,
`source`, `Achilles_n_replicates`, `cell_line_NNMD`, `culture_type`,
`culture_medium`, and `cas9_activity`.
}
\details{
This data represents the `depmap_19Q1_cell_lines` file taken from
the 19Q1 [Broad Institute](https://depmap.org/portal/download/)
cancer depenedency study. This dataset features the a primary key `depmap_id`
which is a unique ID given to each cell line and is found in the first column
of this dataset. The `depmap_id` attribute is used as a foreign key in all
other datasets in the package. This dataset has been converted to a long
format tibble. This dataset does not contain any expression or dependency
data but rather contains the metadata for all cancer cell lines used in the
depmap project. Variables names were converted to lower case, put in snake
case, and abbreviated where feasible (e.g. "Sanger ID" was changed to
"sanger_id").

This data represents the `sample_info.csv` file taken from the 19Q2
[Broad Institute](https://depmap.org/portal/download/) cancer depenedency
study. This dataset features the a primary key `depmap_id` which is a unique
ID given to each cell line and is found in the first column of this dataset.
The `depmap_id` attribute is used as a foreign key in all other datasets in
the package. This dataset has been converted to a long format tibble. This
dataset does not contain any expression or dependency data but rather
contains the metadata for all cancer cell lines used in the depmap project.
Variables names were converted to lower case, put in snake case, and
abbreviated where feasible (e.g. "Sanger ID" was changed to "sanger_id").
}
\section{Change log}{


- 19Q1: none

 

- 19Q2: depmap metadata dataset contains new variables not found in previous
versions, including: `Achilles_n_replicates`, `cell_line_NNMD`,
`culture_type`, `culture_medium`, and `cas9_activity`.
}

\references{
Tsherniak, A., Vazquez, F., Montgomery, P. G., Weir, B. A.,
Kryukov, G., Cowley, G. S., ... & Meyers, R. M. (2017). Defining a cancer
dependency map. Cell, 170(3), 564-576.
(\href{"https://www.ncbi.nlm.nih.gov/pubmed/28753430"}{PubMed})

Tsherniak, A., Vazquez, F., Montgomery, P. G., Weir, B. A.,
Kryukov, G., Cowley, G. S., ... & Meyers, R. M. (2017). Defining a cancer
dependency map. Cell, 170(3), 564-576.
(\href{"https://www.ncbi.nlm.nih.gov/pubmed/28753430"}{PubMed})
}
\keyword{datasets}
