% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public_19Q1_gene_cn.R
\docType{data}
\name{public_19Q1_gene_cn}
\alias{public_19Q1_gene_cn}
\title{public_19Q1_gene_cn}
\format{A data frame with 1604 rows and 23300 variables. Each variable
column contains copy number for a known gene while each row represents a cell
line:
\describe{
    \item{Gene}{HUGO symbol (e.g. "SAP25") and Entrez ID# (e.g. 100316904)}
}}
\source{
\href{"https://depmap.org/portal/download/"}{DepMap, Broad (2019)}
}
\usage{
data(public_19Q1_gene_cn)
}
\description{
Data from DepMap WES Copy Number Data
}
\details{
Genes 23299
Cell Lines 1604
Primary Diseases 38
Lineages 33
}
\examples{
data(public_19Q1_gene_cn)
A1BG_(1) <- attr(public_19Q1_gene_cn, "A1BG (1)")

\dontrun{
public_19Q1_gene_cn
}

}
\references{
Tsherniak, A., Vazquez, F., Montgomery, P. G., Weir, B. A.,
Kryukov, G., Cowley, G. S., ... & Meyers, R. M. (2017). Defining a cancer
dependency map. Cell, 170(3), 564-576.
(\href{"https://www.ncbi.nlm.nih.gov/pubmed/28753430"}{PubMed})
}
\keyword{datasets}
