% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crispr.R
\docType{data}
\name{crispr}
\alias{crispr}
\title{crispr}
\format{A long format tibble data frame with 9839772 rows and 6 columns:
\describe{
    \item{depmap_id}{cell line foreign key (i.e. "ACH-00001")}
    \item{cell_line}{Name of cancer cell line (i.e. "184A1_BREAST")}
    \item{gene}{HUGO symbol (e.g. "SAP25") and Entrez ID# (e.g. 100316904)}
    \item{gene_name}{HUGO symbol (e.g. "SAP25")}
    \item{entrez_id}{Entrez ID# (e.g. 100316904}
    \item{dependency}{numerical depenency score of a gene for a cell line}
}}
\source{
\href{"https://ndownloader.figshare.com/files/14221385"}{DepMap, 
Broad (2019)}
}
\usage{
data(crispr)
}
\description{
The `crispr` dataset is the genetic dependency of CRISPR-Cas9 gene knockout 
of select genes in various cancer cell lines. This dataset includes data
from 17634 genes, 558 cell lines, 26 primary diseases, 28 lineages. The 
columns of `crispr` are: `depmap_id` a foreign key corresponding to the 
cancer cell lineage, `cell_line` containing the common CCLE name of the 
cancer cell lines, `gene` containing both the HUGO gene name of the knockout 
gene along with entrez ID#, `gene_name` containing only the HUGO gene name, 
`entrez_id` containing only the entrez ID# and `dependency` which contains 
the numerical dependency score values for each pair of genes and cell lines.
}
\details{
This data represents the `gene_effect_corrected` file taken from 
the 19Q1 [Broad Institute](https://depmap.org/portal/download/) 
cancer depenedency study. This dataset features the addition of a foreign key
`depmap_id` found in the first column of this dataset, which was added from 
the `metadata` dataset. The numeric column `depenency` contains dependency
scores for a knock out of a known gene for a given cell line. The dataset has
been converted to a long format tibble. Variables names from the original
dataset were converted to lower case, put in snake case, and abbreviated 
where feasible.
}
\examples{
data(crispr)
crispr
}
\references{
Tsherniak, A., Vazquez, F., Montgomery, P. G., Weir, B. A.,
Kryukov, G., Cowley, G. S., ... & Meyers, R. M. (2017). Defining a cancer
dependency map. Cell, 170(3), 564-576.
(\href{"https://www.ncbi.nlm.nih.gov/pubmed/28753430"}{PubMed})
}
\keyword{datasets}
