% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\docType{data}
\name{metadata}
\alias{metadata}
\title{metadata}
\format{A data frame with 1677 rows (cell lines) and 9 variables:
\describe{
  \item{depmap_id}{cell line primary key, used in other datasets as foreign
  key (i.e. "ACH-00001")}
  \item{cell_line}{Name of cancer cell line (i.e. "184A1_BREAST")}
  \item{aliases}{Aliases of cancer cell line}
  \item{cosmic_id}{Catalogue Of Somatic Mutations In Cancer (COSMIC) ID
  number (e.g. 905933)}
  \item{sanger_id}{Sanger ID (eg. 2201)}
  \item{primary_disease}{Primary Disease (e.g. cancer type)}
  \item{subtype_disease}{Subtype Disease (e.g. Acute Myelogenous Leukemia
  (AML), M3 (Promyelocytic))}
  \item{gender}{Gender of patient of cell source}
  \item{source}{Organization where the cell was sampled}
}}
\source{
\href{"https://depmap.org/portal/download/"}{DepMap, Broad (2019)}
}
\usage{
data(metadata)
}
\description{
The `metadata` dataset contains the metadata about cell lines in the 19Q1
Broad Institute DepMap release, and corresponds with the
`depmap_19Q1_cell_lines.csv` file. The metadata includes mapping between
`depmap_id` and `cell_line` name for cancer cell lines. This dataset
contains: 0 Genes, 1677 Cell Lines, 38 Primary Diseases, 33 Lineages. The
columns found in this dataset are `depmap_id`, `cell_line`, `aliases`,
`cosmic_id`, `sanger_id`, `primary_disease`, `subtype_disease`, `gender` and
`source`.
}
\details{
This data represents the `depmap_19Q1_cell_lines` dataset taken from
the public depmap portal. This dataset does not contain any expression or
dependency data but rather contains the metadata for all cancer cell lines
used in the depmap project. `metadata` also contains primary key `depmap_id`
which is used as a foreign key in all other datasets in the package.
}
\examples{
data(metadata)
metadata
}
\references{
Tsherniak, A., Vazquez, F., Montgomery, P. G., Weir, B. A.,
Kryukov, G., Cowley, G. S., ... & Meyers, R. M. (2017). Defining a cancer
dependency map. Cell, 170(3), 564-576.
(\href{"https://www.ncbi.nlm.nih.gov/pubmed/28753430"}{PubMed})
}
\keyword{datasets}
