% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnai.R
\docType{data}
\name{rnai}
\alias{rnai}
\title{rnai}
\format{A data frame with 12324008 rows (each corresponding to an attribute 
of a given cell lines) and 6 columns:
\describe{
    \item{depmap_id}{cell line foreign key (i.e. "ACH-00001")}
    \item{cell_line}{Name of cancer cell line (i.e. "184A1_BREAST")}
    \item{gene}{HUGO symbol (e.g. "SAP25") and Entrez ID# (e.g. 100316904)}
    \item{gene_name}{HUGO symbol (e.g. "SAP25")}
    \item{entrez_id}{Entrez ID# (e.g. 100316904}
    \item{dependency}{numerical depenency score of a gene for a cell line}
}}
\source{
\href{"https://depmap.org/portal/download/"}{DepMap, Broad (2019)}
}
\usage{
data(rnai)
}
\description{
Genetic Dependency Combined RNAi dataset (Broad, Novartis, Marcotte) 
originating from the `D2_combined_genetic_dependency_scores` file and has 
been converted from wide to the long form dataset. Data from RNAi dataset 
includes 17309 genes, 711 cell Lines 711, 30 primary Diseases and 31 
lineages. The first column `depmap_id` is a foreign key relating to the cell 
lineage, followed by `cell_line` which contains the common CCLE name of the 
cancer cell lines, `gene` which contains both the HUGO gene name of the 
knockdown gene along with entrez ID#, followed by the column `gene_name` 
which only contains HUGO gene name and `entrez_id` which contains only the 
entrez ID#. The last column contains the numerical dependency score values 
for each pair of genes and cell lines.
}
\details{
This data represents the `D2_combined_genetic_dependency_scores` 
dataset taken from the public depmap portal. This dataset features the 
addition of a foreign key `depmapID`, which was added from the `metadata` 
dataset. The dataset has been converted to the long format.
}
\examples{
data(rnai)
rnai
}
\references{
Tsherniak, A., Vazquez, F., Montgomery, P. G., Weir, B. A.,
Kryukov, G., Cowley, G. S., ... & Meyers, R. M. (2017). Defining a cancer
dependency map. Cell, 170(3), 564-576.
(\href{"https://www.ncbi.nlm.nih.gov/pubmed/28753430"}{PubMed})
}
\keyword{datasets}
