% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RPPA.R
\docType{data}
\name{RPPA}
\alias{RPPA}
\title{RPPA}
\format{A data frame with 899 rows (cell lines) and 216 variables (genes):
\describe{
    \item{depmap_id}{cell line foreign key (i.e. "ACH-000001")}
    \item{cell_line}{Name of cancer cell line (i.e. "NIHOVCAR3_OVARY")}
    \item{antibody}{Name of antibody targeting protein (i.e. "14-3-3_beta")}
    \item{expression}{protein expression}
}}
\source{
\href{"https://depmap.org/portal/download/"}{DepMap, Broad (2019)}
}
\usage{
data(RPPA)
}
\description{
Celluar Models dataset: CCLE Reverse Phase Protein Array (RPPA) data which 
originates from the `CCLE_RPPA_20180123`. This dataset has been converted 
from wide fromat to the long form dataset and contains 214 Genes, 899 Cell 
Lines, 28 Primary Diseases, 28 Lineages. The first column is the `depmap_id` 
is a foreign key relating to the cell lineage, followed by `cell_line` which 
contains the common CCLE name of the cancer cell lines, `gene` which contains
the knockdown gene expression, `antibody` containing the name of knocked down 
gene and `expression` containing numerical protein expression data.
}
\details{
This data represents the `CCLE_RPPA_20180123` dataset taken from
the public depmap portal. This dataset features the addition of a foreign key
`depmap_id`` found in the first column of this dataset, which was added from 
the `metadata` dataset. The dataset has been converted to the long format.
}
\examples{
data(RPPA)
RPPA
}
\references{
Tsherniak, A., Vazquez, F., Montgomery, P. G., Weir, B. A.,
Kryukov, G., Cowley, G. S., ... & Meyers, R. M. (2017). Defining a cancer
dependency map. Cell, 170(3), 564-576.
(\href{"https://www.ncbi.nlm.nih.gov/pubmed/28753430"}{PubMed})
}
\keyword{datasets}
